shiva/requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
Pygments==2.18.0
pylint==3.3.1
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
